fix -  corrected call chain that optained the begin and end iterators of vectors containing coordinates of block with id in lanes 77 and 118

+++ Minecraft_Block_Finder/Minecraft_Block_Finder/Finder/src/RegionBlockFinder.cpp
@@ -79,7 +79,7 @@ std::vector<Coordinate> RegionBlockFinder::workerSingleThread(size_t chunkX, siz
 
 	size_t nbBlocks = blocksData.size();
 
-	for (size_t index = 0; index < nbBlocks; ++index) {
+	for (int index = 0; index < nbBlocks; ++index) {
 		if (blocksData[index] = m_id) {
 			int sectionX = index % 16;
 			int sectionZ = (index / 16) % 16;
